+++ Big Neon/Screens/GuestListViewController/EventViewController.Swipe.swift
@@ -59,36 +59,6 @@ extension EventViewController: SwipeActionTransitioning {
         }
     }
     
-    
-    //  Reloading Guests Cells
-    func reloadGuestCells(atIndexPath indexPath: IndexPath?) {
-        guard let indexPath = indexPath else { return }
-        let guestCell: EventGuestsCell = self.guestTableView.cellForRow(at: indexPath) as! EventGuestsCell
-        guestCell.ticketStateView.stopAnimation(animationStyle: .normal, revertAfterDelay: 0.0) {
-            guestCell.ticketStateView.layer.cornerRadius = 3.0
-            guestCell.ticketStateView.setTitle("REDEEMED", for: UIControl.State.normal)
-            guestCell.ticketStateView.backgroundColor = UIColor.brandBlack
-        }
-        
-        let guestValues = self.isSearching == true ? self.eventViewModel.guestCoreDataSearchResults : self.eventViewModel.guestCoreData
-        guestValues[indexPath.row].status = TicketStatus.Redeemed.rawValue
-    }
-    
-    //  Checkin Guest
-    func checkinAutomatically(withTicketID ticketID: String, fromGuestTableView: Bool, atIndexPath indexPath: IndexPath?) {
-        self.scannerViewModel.automaticallyCheckin(ticketID: ticketID, eventID: nil) { [weak self] (scanFeedback, errorString, ticket) in
-            DispatchQueue.main.async {
-                if self?.isSearching == true {
-                    self?.eventViewModel.guestCoreDataSearchResults.first(where: { $0.id == ticketID})?.status = TicketStatus.Redeemed.rawValue
-                    self?.reloadGuestCells(atIndexPath: indexPath)
-                } else {
-                    self?.eventViewModel.guestCoreData.first(where: { $0.id == ticketID})?.status = TicketStatus.Redeemed.rawValue
-                    self?.reloadGuestCells(atIndexPath: indexPath)
-                }
-            }
-        }
-    }
-    
     func tableView(_ tableView: UITableView, editActionsForRowAt indexPath: IndexPath, for orientation: SwipeActionsOrientation) -> [SwipeAction]? {
         let guestValues = self.isSearching == true ? self.eventViewModel.guestCoreDataSearchResults : self.eventViewModel.guestCoreData
         return [swipeCellAction(forGuestValues: guestValues, indexPath)]
